- refactor app theming
