+++ lib/core/routing/app_router.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:gem_store_app/core/routing/routes.dart';
 import 'package:gem_store_app/features/cart/cart_screen.dart';
 import 'package:gem_store_app/features/checkout/check_out_done_screen.dart';
-import 'package:gem_store_app/features/onboarding/onboarding_screen.dart';
 
 import '../../features/checkout/checkout_screen.dart';
+import '../../features/onboarding/views/onboarding_screen.dart';
 
 class AppRouter {
   Route? generateRoute(RouteSettings settings) {
