--- lib/core/themes/colors.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract class ColorManager {
-  static const Color mainGreyColor = Color(0xFF464447);
-
-  static const Color lightGreyColor = Color(0xFFE7E8E9);
-
-  static const Color deepGreenColor = Color(0xFF508A7B);
-}
