--- lib/core/theming/app_colors.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract class AppColors {
-  static const primaryColor = Color(0xFF1D1F22);
-  static const scaffoldBackGroundColor = Color(0xFFFAFAFA);
-  static const darkBlueColor = Color(0xFF1E3354);
-  static const whiteColor = Color(0xFFFFFFFF);
-  static const lighterGrayColor = Color(0xFFF1F2F3);
-  static const moreLighterGrayColor = Color(0xFFFDFDFD);
-  static const lightGrayColor = Color(0xFFC8C7CC);
-  static const darkGrayColor = Color(0xFF777E90);
-  static const darkerGrayColor = Color(0xFF343434);
-  static const greenColor = Color(0xFF508A7B);
-
-}
