--- lib/core/untils/app_assets.dart
@@ -1,13 +1,0 @@
-abstract class AppAssets {
-  static String imagesOboardeingStart = "assets/images/onboarding_start.png";
-  static List<String> imagesOboardeingItems = [
-    "assets/images/onboarding_item1.png",
-    "assets/images/onboarding_item2.png",
-    "assets/images/onboarding_item3.png"
-  ];
-
-  static String imagesBackButtonIcon = "assets/images/back_button_icon.png";
-  static String imagesActiveFavoriteIcon =
-      "assets/images/activate_favorite_icon.png";
-  static String imagesTestItem = "assets/images/test_item.png";
-}
