+++ lib/core/utils/app_assets.dart
@@ -23,4 +23,15 @@ class AppAssets {
   static String appleLogo = 'assets/images/appleLogo.png';
   static String googleLogo = 'assets/images/googleLogo.png';
   static String facebookLogo = 'assets/images/facebookLogo.png';
+  static String imagesOboardeingStart = "assets/images/onboarding_start.png";
+  static List<String> imagesOboardeingItems = [
+    "assets/images/onboarding_item1.png",
+    "assets/images/onboarding_item2.png",
+    "assets/images/onboarding_item3.png"
+  ];
+
+  static String imagesBackButtonIcon = "assets/images/back_button_icon.png";
+  static String imagesActiveFavoriteIcon =
+      "assets/images/activate_favorite_icon.png";
+  static String imagesTestItem = "assets/images/test_item.png";
 }
