--- lib/core/utils/assets_manager.dart
@@ -1,4 +1,0 @@
-class SVGAssets{
-  static const String _basePath = "assets/svg/";
-  static const String filterIcon = "${_basePath}Filter_big.svg";
-}
