--- lib/core/utils/color_manager.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ColorManager {
-  ColorManager._();
-
-  static const Color primary = Color(0xff33302E);
-  static const Color black = Color(0xff000000);
-  static const Color black33 = Color(0xff33302E);
-  static const Color transparent = Colors.transparent;
-}
