--- lib/core/utils/string_manager.dart
@@ -1,3 +1,0 @@
-class StringsManager {
-
-}
