+++ lib/core/widgets/app_back_button.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../theming/app_colors.dart';
+import '../utils/app_colors.dart';
 
 class AppBackButton extends StatelessWidget {
   final VoidCallback? onPressed;
@@ -24,7 +24,7 @@ class AppBackButton extends StatelessWidget {
           decoration: BoxDecoration(
             boxShadow: [
               BoxShadow(
-                color: AppColors.primaryColor.withOpacity(0.2),
+                color: AppColors.blackColor.withOpacity(0.2),
                 blurRadius: 10.r,
                 offset: const Offset(0, 2),
               ),
