+++ lib/core/widgets/app_custom_app_bar.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:gem_store_app/core/theming/app_colors.dart';
-import 'package:gem_store_app/core/theming/app_styles.dart';
 import 'package:gem_store_app/core/widgets/app_back_button.dart';
+
+import '../utils/app_colors.dart';
+import '../utils/app_text_styles.dart';
 
 class AppCustomAppBar extends StatelessWidget implements PreferredSizeWidget {
   final String title;
@@ -18,7 +19,7 @@ class AppCustomAppBar extends StatelessWidget implements PreferredSizeWidget {
       leadingWidth: 100.w,
       title: Text(
         title,
-        style: AppStyles.font18BlackSemiBold,
+        style: AppTextStyles.font18BlackSemiBold,
       ),
       centerTitle: true,
     );
