+++ lib/features/cart/cart_screen.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:gem_store_app/core/theming/app_styles.dart';
 import 'package:gem_store_app/core/utils/spacing.dart';
 import 'package:gem_store_app/core/widgets/app_custom_button.dart';
-import 'package:gem_store_app/core/widgets/custom_app_bar.dart';
 import 'package:gem_store_app/features/cart/widgets/cart_list_view.dart';
-import 'package:gem_store_app/features/cart/widgets/cart_price_item.dart';
 import 'package:gem_store_app/features/cart/widgets/cart_price_items_widget.dart';
 
 import '../../core/widgets/app_custom_app_bar.dart';
