+++ lib/features/checkout/checkout_screen.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:gem_store_app/core/theming/app_styles.dart';
 import 'package:gem_store_app/core/utils/spacing.dart';
-import 'package:gem_store_app/core/widgets/custom_app_bar.dart';
 import 'package:gem_store_app/features/checkout/widgets/checkout_inputs.dart';
-import 'package:gem_store_app/features/checkout/widgets/country_drop_down_widget.dart';
 import 'package:gem_store_app/features/checkout/widgets/delivery_home_container.dart';
 import 'package:gem_store_app/features/checkout/widgets/location_and_done_shape.dart';
 
+import '../../core/utils/app_text_styles.dart';
 import '../../core/widgets/app_custom_app_bar.dart';
 import '../../core/widgets/app_custom_button.dart';
 import 'check_out_done_screen.dart';
@@ -29,18 +27,18 @@ class CheckoutScreen extends StatelessWidget {
                 LocationAndDoneShape(),
                 Text(
                   "STEP 1",
-                  style: AppStyles.font11BlackLight,
+                  style: AppTextStyles.font11BlackLight,
                 ),
                 Text(
                   "Shipping",
-                  style: AppStyles.font25BlackBold,
+                  style: AppTextStyles.font25BlackBold,
                 ),
                 verticalSpace(37),
                 CheckoutInputs(),
                 verticalSpace(61),
                 Text(
                   "Shipping method",
-                  style: AppStyles.font25BlackBold,
+                  style: AppTextStyles.font25BlackBold,
                 ),
                 verticalSpace(21),
                 DeliveryHomeContainer(),
