+++ lib/features/checkout/widgets/country_drop_down_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:gem_store_app/core/theming/app_colors.dart';
+
+import '../../../core/utils/app_colors.dart';
 
 class CountryDropdownWidget extends StatefulWidget {
   const CountryDropdownWidget({super.key});
