+++ lib/features/discover/presentation/views/widgets/custom_searched_item.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:gem_store_app/core/utils/app_assets.dart';
 import 'package:gem_store_app/core/utils/app_colors.dart';
-import 'package:gem_store_app/core/utils/app_text_styles.dart';
+
+import '../../../../../core/utils/app_text_styles.dart';
 
 class CustomSearchedItem extends StatelessWidget {
   const CustomSearchedItem({super.key});
