+++ lib/features/onboarding/views/widgets/scrolling_items_and_Indicator.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:gem_store_app/core/themes/colors.dart';
+import 'package:gem_store_app/core/utils/app_colors.dart';
 import 'package:gem_store_app/features/onboarding/cubits/scrolling/scrolling_cubit.dart';
 import 'package:gem_store_app/features/onboarding/cubits/scrolling/scrolling_state.dart';
 import 'package:gem_store_app/features/onboarding/views/widgets/srolling_item.dart';
 
-import '../../../../core/untils/app_assets.dart';
+import '../../../../core/utils/app_assets.dart';
+
 
 class ScrollingItemsAndIndicator extends StatelessWidget {
   const ScrollingItemsAndIndicator({super.key});
@@ -52,7 +53,7 @@ class ScrollingItemsAndIndicator extends StatelessWidget {
                     height: 7.h,
                     width: 7.w,
                     decoration: BoxDecoration(
-                      color: index == currentPage ? Colors.white : ColorManager.mainGreyColor,
+                      color: index == currentPage ? Colors.white : AppColors.greyContainerColor,
                       borderRadius: BorderRadius.circular(10),
                       border: Border.all(color: Colors.white),
                     ),
