+++ lib/features/onboarding/views/widgets/srolling_item.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../core/untils/app_assets.dart';
+
+import '../../../../core/utils/app_assets.dart';
+
 
 class CustomeItem extends StatelessWidget {
   const CustomeItem({
