+++ lib/features/track_order/track_order.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:gem_store_app/core/theming/app_colors.dart';
-import 'package:gem_store_app/core/theming/app_styles.dart';
 import 'package:gem_store_app/core/utils/app_text_styles.dart';
 import 'package:gem_store_app/core/utils/spacing.dart';
 import 'package:gem_store_app/core/widgets/app_custom_app_bar.dart';
 import 'package:gem_store_app/features/track_order/widgets/track_order_list_view.dart';
+
+import '../../core/utils/app_colors.dart';
 
 class TrackOrderScreen extends StatelessWidget {
   const TrackOrderScreen({super.key});
@@ -28,8 +28,8 @@ class TrackOrderScreen extends StatelessWidget {
               children: [
                 TextSpan(
                   text: " 15.05.21",
-                  style: AppStyles.font14DarkGrayMedium.copyWith(
-                    color: AppColors.primaryColor,
+                  style: AppTextStyles.font14DarkGrayMedium.copyWith(
+                    color: AppColors.blackColor,
                   ),
                 ),
               ],
@@ -43,8 +43,8 @@ class TrackOrderScreen extends StatelessWidget {
                 children: [
                   TextSpan(
                     text: " IK287368838",
-                    style: AppStyles.font14DarkGrayMedium.copyWith(
-                      color: AppColors.primaryColor,
+                    style: AppTextStyles.font14DarkGrayMedium.copyWith(
+                      color: AppColors.blackColor,
                     ),
                   ),
                 ],
